Update fedora virt install setup

+++ tools/virt-install/setup/fedora.sh
@@ -24,9 +24,11 @@ sed -i 's/^GRUB_TIMEOUT=.*/GRUB_TIMEOUT=0/g' /etc/default/grub
 grub2-mkconfig -o /boot/grub2/grub.cfg
 
 # cloud-init fix
-wget https://dl.fedoraproject.org/pub/fedora/linux/development/rawhide/Everything/x86_64/os/Packages/c/cloud-init-25.3-1.fc44.noarch.rpm
-echo "51c33d2b8780480fa1aa482824e7850e21363e2713a8197e2ed97a8d28c14adb  cloud-init-25.3-1.fc44.noarch.rpm" | dnf -y install cloud-init-25.3-1.fc44.noarch.rpm
-rm -f cloud-init-25.3-1.fc44.noarch.rpm
+if [[ "$(cloud-init --version 2>&1)" == *"25.2"* ]]; then
+  wget https://dl.fedoraproject.org/pub/fedora/linux/development/rawhide/Everything/x86_64/os/Packages/c/cloud-init-25.3-1.fc44.noarch.rpm
+  echo "51c33d2b8780480fa1aa482824e7850e21363e2713a8197e2ed97a8d28c14adb  cloud-init-25.3-1.fc44.noarch.rpm" | dnf -y install cloud-init-25.3-1.fc44.noarch.rpm
+  rm -f cloud-init-25.3-1.fc44.noarch.rpm
+fi
 
 systemctl daemon-reload
 systemctl enable qemu-guest-agent.service
